fix assesment
